agri_python/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.6
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.1
drf-yasg==1.21.8
